디렉토리 구조 변경
home, profile 추가
