+++ app/src/main/AndroidManifest.xml
@@ -67,7 +67,7 @@
             android:name=".member.view.FindIdActivity"
             android:exported="false" />
         <activity
-            android:name=".board.view.activity.HomeActivity"
+            android:name=".home.view.HomeActivity"
             android:exported="false" />
         <activity
             android:name=".util.view.MainActivity"
