+++ app/src/main/java/com/team1/teamone/util/view/SplashActivity.kt
@@ -5,9 +5,7 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.Handler
 import com.team1.teamone.R
-import com.team1.teamone.board.view.activity.WriteFreeBoardActivity
-import com.team1.teamone.board.view.activity.WriteRecruitmentBoardActivity
-import com.team1.teamone.member.view.activity.LoginActivity
+import com.team1.teamone.home.view.HomeActivity
 
 class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -16,7 +14,7 @@ class SplashActivity : AppCompatActivity() {
 
 
         Handler().postDelayed({
-            val intent = Intent(this, LoginActivity::class.java)
+            val intent = Intent(this, HomeActivity::class.java)
             startActivity(intent)
             finish()
         }, 3000)
