+++ app/src/main/res/layout/activity_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".board.view.activity.HomeActivity">
+    tools:context=".home.view.HomeActivity">
 
 <!--    FragementContainerView 삭제-->
 <!--    <androidx.fragment.app.FragmentContainerView-->
