+++ app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".board.view.fragment.HomeFragment">
+        tools:context=".home.view.HomeFragment">
 
         <androidx.coordinatorlayout.widget.CoordinatorLayout
             android:layout_width="match_parent"
