+++ app/src/main/res/layout/fragment_message.xml
@@ -5,7 +5,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".board.view.fragment.MessageFragment">
+        tools:context=".message.MessageFragment">
 
         <TextView
             android:layout_width="wrap_content"
