+++ app/src/main/res/layout/fragment_profile.xml
@@ -5,7 +5,7 @@
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".board.view.fragment.ProfileFragment">
+        tools:context=".profile.ProfileFragment">
 
         <TextView
             android:text="Profile"
