+++ app/src/main/res/navigation/main_nav.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.team1.teamone.board.view.fragment.HomeFragment"
+        android:name="com.team1.teamone.home.view.HomeFragment"
         android:label="fragment_book_mark"
         tools:layout="@layout/fragment_home" >
         <action
@@ -22,7 +22,7 @@
     </fragment>
     <fragment
         android:id="@+id/messageFragment"
-        android:name="com.team1.teamone.board.view.fragment.MessageFragment"
+        android:name="com.team1.teamone.message.MessageFragment"
         android:label="fragment_message"
         tools:layout="@layout/fragment_message" >
         <action
@@ -37,7 +37,7 @@
     </fragment>
     <fragment
         android:id="@+id/profileFragment"
-        android:name="com.team1.teamone.board.view.fragment.ProfileFragment"
+        android:name="com.team1.teamone.profile.ProfileFragment"
         android:label="fragment_profile"
         tools:layout="@layout/fragment_profile" >
         <action
